% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alignBowtie}
\alias{alignBowtie}
\title{aligns reads using bowtie}
\usage{
alignBowtie(fastq, output, bowtiepath, bowtieref, shortreads, verbose = TRUE)
}
\arguments{
\item{fastq}{full path to fastq file}

\item{bowtiepath}{full path to bowtie}

\item{bowtieref}{full path to bowtie reference}

\item{arguments}{arguments to give to bowtie}

\item{verbose}{boolean whether or not to print command}

\item{nlines}{the number of lines to look at to determine the scoring method}
}
\description{
This function aligns reads using bowtie
}

