% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dose.R
\name{theme_dose}
\alias{theme_dose}
\title{MANGO default plotting theme}
\usage{
theme_dose(base_size = 12)
}
\arguments{
\item{base_size}{Base font size used by the theme.}
}
\value{
A ggplot2 theme object.
}
\description{
A ggplot2 theme used across MANGO plotting functions (grid-on style).
}
