% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignBowtie.R
\name{alignBowtie}
\alias{alignBowtie}
\title{aligns reads using bowtie}
\usage{
alignBowtie(fastq, output, bowtiepath, bowtieref, shortreads, threads,
  verbose = TRUE, nlines = 10000)
}
\arguments{
\item{fastq}{full path to fastq file}

\item{bowtiepath}{full path to bowtie}

\item{bowtieref}{full path to bowtie reference}

\item{threads}{the number of threads to be used for bowtie alignment}

\item{verbose}{boolean whether or not to print command}

\item{nlines}{the number of lines to look at to determine the scoring method}

\item{arguments}{arguments to give to bowtie}
}
\description{
This function aligns reads using bowtie
}

